quiz_bot/requirements.txt
python-telegram-bot>=20.0,<21.0
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
python-dotenv>=1.0.0  # opzionale ma utile per caricare variabili da .env
